+++ src/main/scala/nl/knaw/dans/easy/pid/generator/DatabaseComponent.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid.seedstorage
+package nl.knaw.dans.easy.pid.generator
 
 import java.sql.{ Connection, SQLException }
